fix: miscellaneous (#12)
* chore: use cast wallet, move chain name from env to script

* chore: add slipstream, add NONE connector, deploy

+++ contracts/interfaces/IComet.sol
@@ -5,6 +5,6 @@ pragma solidity 0.8.23;
 // CompoundV3 money market interface
 interface IComet {
     function baseToken() external view returns (address);
-    function supply(address asset, uint amount) external;
-    function withdraw(address asset, uint amount) external;
+    function supply(address asset, uint256 amount) external;
+    function withdraw(address asset, uint256 amount) external;
 }
